Duplicate data transformation

+++ src/screens/models/configure/configuration.ts
@@ -302,6 +302,16 @@ export class ModelsConfigureConfiguration extends connect(store)(PageViewElement
                 </td>
             </tr>
 
+            <tr>
+                <td>Usage notes:</td>
+                <td>
+                    ${this._editing ? html`
+                    <textarea id="form-config-usage-notes" rows="6">${this._config.hasUsageNotes}</textarea>
+                ` : this._config.hasUsageNotes}
+                </td>
+            </tr>
+
+
             ${this._editing ? html`
             <tr>
                 <td>Tag</td>
@@ -359,10 +369,11 @@ export class ModelsConfigureConfiguration extends connect(store)(PageViewElement
     }
 
     private _onDeleteButtonClicked () {
-        //TODO
-        store.dispatch(modelConfigurationDelete( this._config ));
-        this._scrollUp();
-        goToPage('models/configure/');
+        if (confirm('This configuration and all its associated resources (variables, files) will be deleted. Are you sure?')) {
+            store.dispatch(modelConfigurationDelete( this._config ));
+            this._scrollUp();
+            goToPage('models/configure/');
+        }
     }
 
     private _onCancelButtonClicked () {
@@ -382,6 +393,7 @@ export class ModelsConfigureConfiguration extends connect(store)(PageViewElement
         let inputWebsite : Textfield = this.shadowRoot.getElementById("form-config-website") as Textfield;
         let inputCompLoc : HTMLTextAreaElement = this.shadowRoot.getElementById("form-config-comp-loc") as HTMLTextAreaElement;
         let inputTag : Select = this.shadowRoot.getElementById("form-config-tag") as Select;
+        let inputNote : Textfield = this.shadowRoot.getElementById("form-config-usage-notes") as Textfield;
 
         let name        : string = inputName        ? inputName        .value : '';
         let category    : string = inputCategory    ? inputCategory    .value : '';
@@ -393,6 +405,7 @@ export class ModelsConfigureConfiguration extends connect(store)(PageViewElement
         let website     : string = inputWebsite     ? inputWebsite     .value : '';
         let compLoc     : string = inputCompLoc     ? inputCompLoc     .value : '';
         let tag         : string = inputTag         ? inputTag         .value : '';
+        let notes : string = inputNote ? inputNote.value : '';
 
         if (name && category && desc) {
             let jsonObj = {
@@ -417,6 +430,7 @@ export class ModelsConfigureConfiguration extends connect(store)(PageViewElement
             if (website) jsonObj['website'] = [website];
             if (compLoc) jsonObj['hasComponentLocation'] = [compLoc];
             if (tag) jsonObj['tag'] = [tag];
+            if (notes) jsonObj['hasUsageNotes'] = [notes];
 
             // save parameters first
             let promises = [];
